+++ changelog.md
@@ -6,6 +6,7 @@
 > This version is not released yet and is under active development.
 
 - Switch from from `requirements/*.txt` files to hard-coding version dependencies for workflow tools.
+- Check consistency between project and repository metadata.
 
 ## [5.0.1 (2026-01-22)](https://github.com/kdeldycke/workflows/compare/v5.0.0...v5.0.1)